Move status printout into 'status', alongside move of style to utils
