+++ src/lib.rs
@@ -3,6 +3,7 @@ mod map;
 pub mod reader;
 pub mod routes;
 pub mod server;
+pub mod utils;
 
 pub use config::{Config, Source};
 pub use server::TileServer;
